ssize_t/size_t vs. int fixes for warning-free compilation on 64-bit arch, from Adriaan de Groot

+++ libpisock/inet.c
@@ -60,8 +60,8 @@ static int pi_inet_bind(pi_socket_t *ps, struct sockaddr *addr,
 static int pi_inet_listen(pi_socket_t *ps, int backlog);
 static int pi_inet_accept(pi_socket_t *ps, struct sockaddr *addr,
 			 size_t *addrlen);
-static int pi_inet_read(pi_socket_t *ps, pi_buffer_t *msg, size_t len, int flags);
-static int pi_inet_write(pi_socket_t *ps, unsigned char *msg, size_t len, int flags);
+static ssize_t pi_inet_read(pi_socket_t *ps, pi_buffer_t *msg, size_t len, int flags);
+static ssize_t pi_inet_write(pi_socket_t *ps, unsigned char *msg, size_t len, int flags);
 static int pi_inet_getsockopt(pi_socket_t *ps, int level, int option_name, 
 			      void *option_value, size_t *option_len);
 static int pi_inet_setsockopt(pi_socket_t *ps, int level, int option_name, 
@@ -499,7 +499,7 @@ pi_inet_accept(pi_socket_t *ps, struct sockaddr *addr, size_t *addrlen)
  * Returns:     number of bytes written or -1 on error
  *
  ***********************************************************************/
-static int
+static ssize_t
 pi_inet_write(pi_socket_t *ps, unsigned char *msg, size_t len, int flags)
 {
 	int 	total,
@@ -545,7 +545,7 @@ pi_inet_write(pi_socket_t *ps, unsigned char *msg, size_t len, int flags)
  * Returns:     number of bytes read or -1 on error
  *
  ***********************************************************************/
-static int
+static ssize_t
 pi_inet_read(pi_socket_t *ps, pi_buffer_t *msg, size_t len, int flags)
 {
 	int 	r,
